+++ appserver/src/objects/units/units.cpp
@@ -11,9 +11,9 @@
 
 namespace Units {
 
-  void Remove_Unit(std::unordered_map<Component::Position, uint8_t> &unitPositions, std::vector<uint8_t> &emptyUnitSlots, int8_t x, int8_t y) {
+  void Remove_Unit(std::unordered_map<Component::Position, int> &unitPositions, std::vector<int> &emptyUnitSlots, int x, int y) {
     Component::Position pos = {x, y};
-    uint8_t index = unitPositions[pos];
+    int index = unitPositions[pos];
     //can't dp this because it will invalidate map indexes
     //    units.erase(units.begin() + index);
     //instead, save the index to a vector of empty slots
@@ -22,7 +22,7 @@ namespace Units {
     unitPositions.erase(pos);
   }
 
-  void Add_Unit(std::vector<Units::Unit> &units, std::unordered_map<Component::Position, uint8_t> &unitPositions, std::vector<uint8_t> &emptyUnitSlots, Units::Unit &unit, int8_t x, int8_t y) {
+  void Add_Unit(std::vector<Units::Unit> &units, std::unordered_map<Component::Position, int> &unitPositions, std::vector<int> &emptyUnitSlots, Units::Unit &unit, int x, int y) {
     if (!emptyUnitSlots.empty()) {
       int index = emptyUnitSlots[emptyUnitSlots.size() - 1];
       emptyUnitSlots.pop_back();
@@ -37,7 +37,7 @@ namespace Units {
     }
   }
 
-  Units::Unit& Get_Unit_At_Position(std::vector<Units::Unit> &units, std::unordered_map<Component::Position, uint8_t> &unitPositions, int8_t x, int8_t y) {
+  Units::Unit& Get_Unit_At_Position(std::vector<Units::Unit> &units, std::unordered_map<Component::Position, int> &unitPositions, int x, int y) {
     if (Get_Unit_Index(unitPositions, x, y) == -1) {
       std::cout << "unit not found at: " << x << " " << y << std::endl;
       return units[0];
@@ -45,7 +45,7 @@ namespace Units {
     return units[Units::Get_Unit_Index(unitPositions, x, y)];
   }
 
-  int8_t Get_Unit_Index(std::unordered_map<Component::Position, uint8_t> &unitPositions, int8_t x, int8_t y) {
+  int Get_Unit_Index(std::unordered_map<Component::Position, int> &unitPositions, int x, int y) {
     Component::Position pos = {x, y};
     if (unitPositions.count(pos) == 1) {
       std::cout << "unit found at: " << x << " " << y << std::endl;
@@ -54,7 +54,7 @@ namespace Units {
     return -1;
   }
 
-  void Update_Unit_Position(std::unordered_map<Component::Position, uint8_t> &unitPositions, const int8_t &x, const int8_t &y, const int8_t &newX, const int8_t &newY) {
+  void Update_Unit_Position(std::unordered_map<Component::Position, int> &unitPositions, const int &x, const int &y, const int &newX, const int &newY) {
     std::cout << "updating unit position from: " << x << " " << y << " to: " << newX << " " << newY << std::endl;
     Component::Position pos = {x, y};
     int index = unitPositions[pos];
@@ -63,7 +63,7 @@ namespace Units {
     unitPositions.emplace(newPos, index);
   }
 
-  void Update_Unit_Position(std::unordered_map<Component::Position, uint8_t> &unitPositions, const Component::Position &pos, const Component::Position &newPos) {
+  void Update_Unit_Position(std::unordered_map<Component::Position, int> &unitPositions, const Component::Position &pos, const Component::Position &newPos) {
     std::cout << "updating unit position from: " << pos.x << " " << pos.y << " to: " << newPos.x << " " << newPos.y << std::endl;
     int index = unitPositions[pos];
     unitPositions.erase(pos);
